# swagger.yaml
openapi: 3.0.1
info:
  title: Rune Airdrop Backend
  description: "Backend implementation Rune Claiming Process"
  version: 0.0.1
servers:
  - url: "http://localhost:5000/api"
  - url: "https://rune-airdrop-backend.onrender.com/api"
...
paths:
...
  /same-amount:
    post:
      summary: This endpoint is used to transfer same amount rune token to different addresses.
      description: This endpoint is used to transfer same amount rune token to different addresses.
      tags:
        - Main
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/sameAmount"
      responses:
        "200":
          description: Same Amount Rune Airdrop Success!
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Same Amount Rune Airdrop failed!
          content:
            application/json:
              schema:
                type: string
...
  /estimate/same-amount:
    post:
      summary: This endpoint is used to estimate same amount rune token to different addresses.
      description: This endpoint is used to estimate same amount rune token to different addresses.
      tags:
        - Estimate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateSameAmount"
      responses:
        "200":
          description: Same Amount Rune Airdrop Estimate Success!
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Same Amount Rune Airdrop Estimate failed!
          content:
            application/json:
              schema:
                type: string
components:
  schemas:
...
    sameAmount:
      type: object
      properties:
        rune_id:
          type: string
          example: "2822267:1023"
        feeRate:
          type: number
          example: 25
        amount:
          type: number
          example: 10
        addressList:
          type: string
          example:
            [
              "tb1pqu4ty7cc065rtl83jxxsyrjtgl4gkyxwx8v448w947ukeszh5lzqmguhu2",
              "tb1pqu4ty7cc065rtl83jxxsyrjtgl4gkyxwx8v448w947ukeszh5lzqmguhu2",
              "tb1pqu4ty7cc065rtl83jxxsyrjtgl4gkyxwx8v448w947ukeszh5lzqmguhu2",
              "tb1pqu4ty7cc065rtl83jxxsyrjtgl4gkyxwx8v448w947ukeszh5lzqmguhu2",
              "tb1pqu4ty7cc065rtl83jxxsyrjtgl4gkyxwx8v448w947ukeszh5lzqmguhu2",
              "tb1pqu4ty7cc065rtl83jxxsyrjtgl4gkyxwx8v448w947ukeszh5lzqmguhu2",
              "tb1pqu4ty7cc065rtl83jxxsyrjtgl4gkyxwx8v448w947ukeszh5lzqmguhu2",
              "tb1pqu4ty7cc065rtl83jxxsyrjtgl4gkyxwx8v448w947ukeszh5lzqmguhu2",
              "tb1pqu4ty7cc065rtl83jxxsyrjtgl4gkyxwx8v448w947ukeszh5lzqmguhu2",
              "tb1pqu4ty7cc065rtl83jxxsyrjtgl4gkyxwx8v448w947ukeszh5lzqmguhu2",
              "tb1pqu4ty7cc065rtl83jxxsyrjtgl4gkyxwx8v448w947ukeszh5lzqmguhu2",
            ]
      required:
        - rune_id
        - feeRate
        - amount
        - addressList
...
    EstimateSameAmount:
      type: object
      properties:
        feeRate:
          type: number
          example: 25
        size:
          type: number
          example: 5000
      required:
        - feeRate
        - size